+++ app/views/intakes/index.html.erb
@@ -1,7 +1,7 @@
 <% content_for :full_page_content do %>
   <%= react_component("BaseContainer", props: {
     userDisplayName: current_user.display_name,
-    currentIntake: @current_intake ? @current_intake.ui_hash : {},
+    serverIntake: @current_intake ? @current_intake.ui_hash : {},
     dropdownUrls: dropdown_urls,
     page: "Intake",
     feedbackUrl: ENV["CASEFLOW_FEEDBACK_URL"],
