+++ client/app/intake/index.js
@@ -1,18 +1,25 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
-import { createStore, applyMiddleware, compose } from 'redux';
+import { createStore, applyMiddleware, combineReducers, compose } from 'redux';
 import perfLogger from 'redux-perf-middleware';
 import IntakeFrame from './IntakeFrame';
-import { reducer, mapDataToInitialState } from './redux/reducer';
+import { intakeReducer, mapDataToInitialIntake } from './reducers/intake';
+import { rampElectionReducer, mapDataToInitialRampElection } from './reducers/rampElection';
 import { getReduxAnalyticsMiddleware } from '../util/getReduxAnalyticsMiddleware';
 
 const Intake = (props) => {
   // eslint-disable-next-line no-underscore-dangle
   const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
   const store = createStore(
-    reducer,
-    mapDataToInitialState(props),
+    combineReducers({
+      intake: intakeReducer,
+      rampElection: rampElectionReducer
+    }),
+    {
+      intake: mapDataToInitialIntake(props),
+      rampElection: mapDataToInitialRampElection(props)
+    },
     composeEnhancers(applyMiddleware(thunk, perfLogger, getReduxAnalyticsMiddleware('intake')))
   );
