+++ client/app/intake/pages/search.jsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 import { doFileNumberSearch, setFileNumberSearch } from '../actions/common';
 import { REQUEST_STATE, PAGE_PATHS, RAMP_INTAKE_STATES } from '../constants';
-import { getRampElectionStatus } from '../redux/selectors';
+import { getRampElectionStatus } from '../helpers/rampElection';
 
 const rampIneligibleInstructions = <div>
   <p>
@@ -86,7 +86,7 @@ class Search extends React.PureComponent {
     const {
       searchErrorCode,
       searchErrorData,
-      rampElectionStatus,
+      rampElection,
       formType
     } = this.props;
 
@@ -94,7 +94,7 @@ class Search extends React.PureComponent {
       return <Redirect to={PAGE_PATHS.BEGIN} />;
     }
 
-    switch (rampElectionStatus) {
+    switch (getRampElectionStatus(rampElection)) {
     case RAMP_INTAKE_STATES.STARTED:
       return <Redirect to={PAGE_PATHS.REVIEW} />;
     case RAMP_INTAKE_STATES.REVIEWED:
@@ -127,13 +127,13 @@ class Search extends React.PureComponent {
 }
 
 export default connect(
-  (state) => ({
-    fileNumberSearchInput: state.inputs.fileNumberSearch,
-    fileNumberSearchRequestStatus: state.requestStatus.fileNumberSearch,
-    rampElectionStatus: getRampElectionStatus(state),
-    searchErrorCode: state.searchErrorCode,
-    searchErrorData: state.searchErrorData,
-    formType: state.formType
+  ({ intake, rampElection }) => ({
+    rampElection,
+    fileNumberSearchInput: intake.fileNumberSearch,
+    fileNumberSearchRequestStatus: intake.requestStatus.fileNumberSearch,
+    searchErrorCode: intake.searchErrorCode,
+    searchErrorData: intake.searchErrorData,
+    formType: intake.formType
   }),
   (dispatch) => bindActionCreators({
     doFileNumberSearch,
